:recycle: put status body in separate component
see #3642
